\name{mmm2-package}
\alias{mmm2-package}
\alias{mmm2-package}
\docType{package}
\title{
Multivariate marginal models with shared regression parameters
}
\description{
Includes a multivariate marginal model for which the formulation permits fitting models with shared and response specific regression parameters at the same time

Asar, O., Ilk, O. (2013). Flexible multivariate marginal models for analyzing multivariate longitudinal data, with applications in R. Submitted.

}
\details{
\tabular{ll}{
Package: \tab mmm2\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2013-29-12\cr
License: \tab GPL (>=2)\cr
}
}
