\name{mmm2-package}
\alias{mmm2-package}
\alias{mmm2-package}
\docType{package}
\title{
Multivariate marginal models with shared regression parameters
}
\description{
Includes a multivariate marginal model for which the formulation permits fitting models with shared and response specific regression parameters at the same time
}
\details{
\tabular{ll}{
Package: \tab mmm2\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2012-11-04\cr
License: \tab GPL (>=2)\cr
}
}
